+++ kitty/glfw.c
@@ -1922,9 +1922,11 @@ toggle_secure_input(PYNOARG) {
 }
 
 static PyObject*
-macos_cycle_through_os_windows(PYNOARG) {
+macos_cycle_through_os_windows(PyObject *self UNUSED, PyObject *backwards) {
 #ifdef __APPLE__
-    cocoa_cycle_through_os_windows();
+    cocoa_cycle_through_os_windows(PyObject_IsTrue(backwards));
+#else
+    (void)backwards;
 #endif
     Py_RETURN_NONE;
 }
@@ -2616,7 +2618,7 @@ static PyMethodDef module_methods[] = {
     METHODB(set_clipboard_data_types, METH_VARARGS),
     METHODB(get_clipboard_mime, METH_VARARGS),
     METHODB(toggle_secure_input, METH_NOARGS),
-    METHODB(macos_cycle_through_os_windows, METH_NOARGS),
+    METHODB(macos_cycle_through_os_windows, METH_O),
     METHODB(get_content_scale_for_window, METH_NOARGS),
     METHODB(ring_bell, METH_VARARGS),
     METHODB(toggle_fullscreen, METH_VARARGS),